requirements.txt
streamlit==1.28.1
oandapyV20==0.7.2
pandas==2.1.3
plotly==5.17.0
python-dotenv==1.0.0
sqlite3